requirements.txt
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
requests==2.27.1
xhtml2pdf~=0.2.5
reportlab==3.6.6